+++ src/app/components/embedded-content-private-share.js
@@ -93,6 +93,7 @@ export default OneproviderEmbeddedContainer.extend(...mixins, {
     'getDatasetsUrl',
     'showSharesIndex',
     'reloadShareList',
+    'reloadCurrentShareRecord',
     // TODO: VFS-11156 This method will be needed to implement shared files links
     // 'getFileGoToUrl',
   ]),